+++ array.go
@@ -21,15 +21,10 @@
 package zap
 
 import (
-	"sync"
 	"time"
 
 	"go.uber.org/zap/zapcore"
 )
-
-var _errArrayElemPool = sync.Pool{New: func() interface{} {
-	return &errArrayElem{}
-}}
 
 // Array constructs a field with the given key and ArrayMarshaler. It provides
 // a flexible, but still type-safe and efficient, way to add array-like types
@@ -321,35 +316,5 @@ func (nums uintptrs) MarshalLogArray(arr zapcore.ArrayEncoder) error {
 	for i := range nums {
 		arr.AppendUintptr(nums[i])
 	}
-	return nil
-}
-
-type errArray []error
-
-func (errs errArray) MarshalLogArray(arr zapcore.ArrayEncoder) error {
-	for i := range errs {
-		if errs[i] == nil {
-			continue
-		}
-		// To represent each error as an object with an "error" attribute and
-		// potentially an "errorVerbose" attribute, we need to wrap it in a
-		// type that implements LogObjectMarshaler. To prevent this from
-		// allocating, pool the wrapper type.
-		elem := _errArrayElemPool.Get().(*errArrayElem)
-		elem.error = errs[i]
-		arr.AppendObject(elem)
-		elem.error = nil
-		_errArrayElemPool.Put(elem)
-	}
-	return nil
-}
-
-type errArrayElem struct {
-	error
-}
-
-func (e *errArrayElem) MarshalLogObject(enc zapcore.ObjectEncoder) error {
-	// Re-use the error field's logic, which supports non-standard error types.
-	Error(e.error).AddTo(enc)
 	return nil
 }
